+++ components/Search/Search.tsx
@@ -3,7 +3,6 @@ import {FC} from "react"
 
 import {useTranslation} from "next-i18next"
 
-
 interface Props {
   placeholder: string
   searchTerm: string
@@ -26,7 +25,7 @@ const Search: FC<Props> = ({placeholder, searchTerm, onSearch}) => {
       <input
         className="w-full flex-1 rounded-md border border-neutral-600 bg-[#202123] px-4 py-3 pr-10 text-[14px] leading-3 text-white"
         type="text"
-        placeholder={t(placeholder) || ""}
+        placeholder={t(placeholder)}
         value={searchTerm}
         onChange={handleSearchChange}
       />