+++ components/Settings/ApiKey.tsx
@@ -13,7 +13,7 @@ interface Props {
   onApiKeyChange: (apiKey: string) => void
 }
 
-export const Key: FC<Props> = ({apiKey, onApiKeyChange}) => {
+export const ApiKey: FC<Props> = ({apiKey, onApiKeyChange}) => {
   const {t} = useTranslation("sidebar")
   const [isChanging, setIsChanging] = useState(false)
   const [newKey, setNewKey] = useState(apiKey)
@@ -48,7 +48,7 @@ export const Key: FC<Props> = ({apiKey, onApiKeyChange}) => {
         value={newKey}
         onChange={(e) => setNewKey(e.target.value)}
         onKeyDown={handleEnterDown}
-        placeholder={t("API key") || "API Key"}
+        placeholder={t("API key")}
       />
 
       <div className="flex w-[40px]">