+++ components/Settings/SettingsDialog.tsx
@@ -16,7 +16,7 @@ interface Props {
   onClose: () => void
 }
 
-export const SettingDialog: FC<Props> = ({open, onClose}) => {
+export const SettingsDialog: FC<Props> = ({open, onClose}) => {
   const {t} = useTranslation("settings")
   const settings: Settings = getSettings()
   const {state, dispatch} = useCreateReducer<Settings>({